Add testsStart script
