+++ bin/testsComplete.js
@@ -2,143 +2,8 @@
 /* eslint-disable no-console */
 
 const admin = require('firebase-admin')
-const isString = require('lodash/isString')
-const fs = require('fs')
-const path = require('path')
-const localTestConfigPath = path.join(__dirname, '..', 'cypress', 'config.json')
-const serviceAccountPath = path.join(__dirname, '..', 'serviceAccount.json')
-const prefixesByCiEnv = {
-  staging: 'STAGE_',
-  production: 'PROD_'
-}
+const utils = require('./utils')
 
-/**
- * Get prefix for current environment based on environment vars available
- * within CI. Falls back to staging (i.e. STAGE)
- * @return {String} Environment prefix string
- */
-function getEnvPrefix() {
-  return (
-    prefixesByCiEnv[process.env.CI_ENVIRONMENT_SLUG] || prefixesByCiEnv.staging
-  )
-}
-
-/**
- * Get environment variable based on the current CI environment
- * @param  {String} varNameRoot - variable name without the environment prefix
- * @return {Any} Value of the environment variable
- * @example
- * envVarBasedOnCIEnv('FIREBASE_PROJECT_ID')
- * // => 'fireadmin-stage' (value of 'STAGE_FIREBASE_PROJECT_ID' environment var)
- */
-function envVarBasedOnCIEnv(varNameRoot) {
-  const prefix = getEnvPrefix()
-  const combined = `${prefix}${varNameRoot}`
-  if (!process.env.CI && !process.env.CI_ENVIRONMENT_SLUG) {
-    const configObj = require(localTestConfigPath)
-    console.log(
-      `Running in local environment, ${
-        configObj[combined] ? combined : varNameRoot
-      } is being loaded from cypress/config.json`
-    )
-    return configObj[combined] || configObj[varNameRoot]
-  }
-  return process.env[combined] || process.env[varNameRoot]
-}
-
-/**
- * Get parsed value of environment variable. Useful for environment variables
- * which have characters that need to be escaped.
- * @param  {String} varNameRoot - variable name without the environment prefix
- * @return {Any} Value of the environment variable
- * @example
- * getParsedEnvVar('FIREBASE_PRIVATE_KEY_ID')
- * // => 'fireadmin-stage' (parsed value of 'STAGE_FIREBASE_PRIVATE_KEY_ID' environment var)
- */
-function getParsedEnvVar(varNameRoot) {
-  const val = envVarBasedOnCIEnv(varNameRoot)
-  const prefix = getEnvPrefix()
-  const combinedVar = `${prefix}${varNameRoot}`
-  if (!val) {
-    console.error(
-      `${combinedVar} not found, make sure it is set within environment vars`
-    )
-  }
-  try {
-    if (isString(val)) {
-      return JSON.parse(val)
-    }
-    return val
-  } catch (err) {
-    console.error(`Error parsing ${combinedVar}`)
-    return val
-  }
-}
-
-/**
- * Get service account from either local file or environment variables
- * @return {Object} Service account object
- */
-function getServiceAccount() {
-  // Check for local service account file (Local dev)
-  if (fs.existsSync(serviceAccountPath)) {
-    console.log('local service account being loaded from ./serviceAccount.json')
-    return require(serviceAccountPath)
-  }
-  console.log(
-    'Service Account file does not exist locally, falling back to environment variables'
-  )
-  // Use environment variables (CI)
-  return {
-    type: 'service_account',
-    project_id: envVarBasedOnCIEnv('FIREBASE_PROJECT_ID'),
-    private_key_id: envVarBasedOnCIEnv('FIREBASE_PRIVATE_KEY_ID'),
-    private_key: getParsedEnvVar('FIREBASE_PRIVATE_KEY'),
-    client_email: envVarBasedOnCIEnv('FIREBASE_CLIENT_EMAIL'),
-    client_id: envVarBasedOnCIEnv('FIREBASE_CLIENT_ID'),
-    auth_uri: 'https://accounts.google.com/o/oauth2/auth',
-    token_uri: 'https://accounts.google.com/o/oauth2/token',
-    auth_provider_x509_cert_url: 'https://www.googleapis.com/oauth2/v1/certs',
-    client_x509_cert_url: envVarBasedOnCIEnv('FIREBASE_CERT_URL')
-  }
-}
-
-let adminInstance
-
-/**
- * Initialize Firebase instance from service account (from either local
- * serviceAccount.json or environment variables)
- * @return {Firebase} Initialized Firebase instance
- */
-function initializeFirebase() {
-  try {
-    // Get service account from local file falling back to environment variables
-    const serviceAccount = getServiceAccount()
-
-    if (!adminInstance) {
-      if (!fs.existsSync(serviceAccountPath)) {
-        const missingAccountErr = `Service account not found, check: ${serviceAccountPath}`
-        console.error(missingAccountErr)
-        throw new Error(missingAccountErr)
-      }
-      console.log(
-        'service account exists, project id: ',
-        serviceAccount.project_id
-      )
-      adminInstance = admin.initializeApp({
-        credential: admin.credential.cert(serviceAccount),
-        databaseURL: `https://${serviceAccount.project_id}.firebaseio.com`
-      })
-      adminInstance.firestore().settings({ timestampsInSnapshots: true })
-    }
-    return adminInstance
-  } catch (err) {
-    console.log(
-      'Error initializing firebase-admin instance from service account.'
-    )
-    throw err
-  }
-}
 var stdin = process.openStdin()
 
 var data = ''
@@ -148,7 +13,7 @@ stdin.on('data', function(chunk) {
 })
 
 stdin.on('end', function() {
-  const db = initializeFirebase().database()
+  const db = utils.initializeFirebase().database()
   if (!process.env.JOB_RUN_KEY) {
     console.log(
       'JOB_RUN_KEY not found within environment, exiting with error status'
@@ -157,10 +22,13 @@ stdin.on('end', function() {
   } else {
     const testMetaPath = `test_runs_meta/${process.env.JOB_RUN_KEY}`
     const resultsRef = db.ref(testMetaPath)
-    const status = data === '0' || data === 0 ? 'success' : 'error'
+    const status = data === '0' || data === 0 ? 'complete' : 'error'
     console(`Writing status "${status}" to ${testMetaPath}`)
     return resultsRef
-      .update({ status })
+      .update({
+        status,
+        completedAt: admin.database.ServerValue.TIMESTAMP
+      })
       .then(() => {
         process.exit()
       })
